v 2.0.9 (#19)

+++ src/main/kotlin/dev/merge/client/hris/models/TimeOffRequest.kt
@@ -34,11 +34,11 @@ import dev.merge.client.shared.ApiClient
  *
  * @param employee The employee requesting time off.
  * @param approver The Merge ID of the employee with the ability to approve the time off request.
- * @param status The status of this time off request.
+ * @param status The status of this time off request.  * `REQUESTED` - REQUESTED * `APPROVED` - APPROVED * `DECLINED` - DECLINED * `CANCELLED` - CANCELLED * `DELETED` - DELETED
  * @param employeeNote The employee note for this time off request.
- * @param units The measurement that the third-party integration uses to count time requested.
+ * @param units The measurement that the third-party integration uses to count time requested.  * `HOURS` - HOURS * `DAYS` - DAYS
  * @param amount The time off quantity measured by the prescribed “units”.
- * @param requestType The type of time off request.
+ * @param requestType The type of time off request.  * `VACATION` - VACATION * `SICK` - SICK * `PERSONAL` - PERSONAL * `JURY_DUTY` - JURY_DUTY * `VOLUNTEER` - VOLUNTEER * `BEREAVEMENT` - BEREAVEMENT
  * @param startTime The day and time of the start of the time requested off.
  * @param endTime The day and time of the end of the time requested off.
  * @param integrationParams 
@@ -56,7 +56,7 @@ data class TimeOffRequest (
     @field:JsonProperty("approver")
     val approver: java.util.UUID? = null,
 
-    /* The status of this time off request. */
+    /* The status of this time off request.  * `REQUESTED` - REQUESTED * `APPROVED` - APPROVED * `DECLINED` - DECLINED * `CANCELLED` - CANCELLED * `DELETED` - DELETED */
     @field:JsonProperty("status")
     val status: TimeOffStatusEnum? = null,
 
@@ -64,15 +64,15 @@ data class TimeOffRequest (
     @field:JsonProperty("employee_note")
     val employeeNote: kotlin.String? = null,
 
-    /* The measurement that the third-party integration uses to count time requested. */
+    /* The measurement that the third-party integration uses to count time requested.  * `HOURS` - HOURS * `DAYS` - DAYS */
     @field:JsonProperty("units")
     val units: UnitsEnum? = null,
 
     /* The time off quantity measured by the prescribed “units”. */
     @field:JsonProperty("amount")
-    val amount: kotlin.Float? = null,
+    val amount: kotlin.Double? = null,
 
-    /* The type of time off request. */
+    /* The type of time off request.  * `VACATION` - VACATION * `SICK` - SICK * `PERSONAL` - PERSONAL * `JURY_DUTY` - JURY_DUTY * `VOLUNTEER` - VOLUNTEER * `BEREAVEMENT` - BEREAVEMENT */
     @field:JsonProperty("request_type")
     val requestType: RequestTypeEnum? = null,
